# yelp-api-openapi.json
{"openapi":"3.1.0","info":{"title":"Yelp","description":"FastAPI","version":"1.0.0"},"paths":{"/api/v1/businesses/":{"get":{"tags":["businesses"],"summary":"Read Businesses","description":"Get all businesses with pagination","operationId":"read_businesses_api_v1_businesses__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Business"},"title":"Response Read Businesses Api V1 Businesses  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/businesses/{business_id}":{"get":{"tags":["businesses"],"summary":"Read Business","description":"Get a specific business by ID","operationId":"read_business_api_v1_businesses__business_id__get","parameters":[{"name":"business_id","in":"path","required":true,"schema":{"type":"string","title":"Business Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Business"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/businesses/city/{city}":{"get":{"tags":["businesses"],"summary":"Read Businesses By City","description":"Get businesses by city","operationId":"read_businesses_by_city_api_v1_businesses_city__city__get","parameters":[{"name":"city","in":"path","required":true,"schema":{"type":"string","title":"City"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Business"},"title":"Response Read Businesses By City Api V1 Businesses City  City  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/businesses/stars/{min_stars}":{"get":{"tags":["businesses"],"summary":"Read Businesses By Stars","description":"Get businesses with minimum star rating","operationId":"read_businesses_by_stars_api_v1_businesses_stars__min_stars__get","parameters":[{"name":"min_stars","in":"path","required":true,"schema":{"type":"number","title":"Min Stars"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Business"},"title":"Response Read Businesses By Stars Api V1 Businesses Stars  Min Stars  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reviews/":{"get":{"tags":["reviews"],"summary":"Read Reviews","description":"Get all reviews with pagination","operationId":"read_reviews_api_v1_reviews__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Review"},"title":"Response Read Reviews Api V1 Reviews  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reviews/{review_id}":{"get":{"tags":["reviews"],"summary":"Read Review","description":"Get a specific review by ID","operationId":"read_review_api_v1_reviews__review_id__get","parameters":[{"name":"review_id","in":"path","required":true,"schema":{"type":"string","title":"Review Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Review"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reviews/business/{business_id}":{"get":{"tags":["reviews"],"summary":"Read Reviews By Business","description":"Get reviews for a specific business","operationId":"read_reviews_by_business_api_v1_reviews_business__business_id__get","parameters":[{"name":"business_id","in":"path","required":true,"schema":{"type":"string","title":"Business Id"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Review"},"title":"Response Read Reviews By Business Api V1 Reviews Business  Business Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reviews/user/{user_id}":{"get":{"tags":["reviews"],"summary":"Read Reviews By User","description":"Get reviews by a specific user","operationId":"read_reviews_by_user_api_v1_reviews_user__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Review"},"title":"Response Read Reviews By User Api V1 Reviews User  User Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/":{"get":{"tags":["users"],"summary":"Read Users","description":"Get all users with pagination","operationId":"read_users_api_v1_users__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"},"title":"Response Read Users Api V1 Users  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/{user_id}":{"get":{"tags":["users"],"summary":"Read User","description":"Get a specific user by ID","operationId":"read_user_api_v1_users__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tips/":{"get":{"tags":["tips"],"summary":"Read Tips","description":"Get all tips with pagination","operationId":"read_tips_api_v1_tips__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Tip"},"title":"Response Read Tips Api V1 Tips  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tips/{user_id}/{business_id}":{"get":{"tags":["tips"],"summary":"Read Tip","description":"Get a specific tip by user and business ID","operationId":"read_tip_api_v1_tips__user_id___business_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"business_id","in":"path","required":true,"schema":{"type":"string","title":"Business Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tip"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tips/business/{business_id}":{"get":{"tags":["tips"],"summary":"Read Tips By Business","description":"Get tips for a specific business","operationId":"read_tips_by_business_api_v1_tips_business__business_id__get","parameters":[{"name":"business_id","in":"path","required":true,"schema":{"type":"string","title":"Business Id"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Tip"},"title":"Response Read Tips By Business Api V1 Tips Business  Business Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tips/user/{user_id}":{"get":{"tags":["tips"],"summary":"Read Tips By User","description":"Get tips by a specific user","operationId":"read_tips_by_user_api_v1_tips_user__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Tip"},"title":"Response Read Tips By User Api V1 Tips User  User Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/checkins/":{"get":{"tags":["checkins"],"summary":"Read Checkins","description":"Get all checkins with pagination","operationId":"read_checkins_api_v1_checkins__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Checkin"},"title":"Response Read Checkins Api V1 Checkins  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/checkins/{business_id}/{date}":{"get":{"tags":["checkins"],"summary":"Read Checkin","description":"Get a specific checkin by business ID and date","operationId":"read_checkin_api_v1_checkins__business_id___date__get","parameters":[{"name":"business_id","in":"path","required":true,"schema":{"type":"string","title":"Business Id"}},{"name":"date","in":"path","required":true,"schema":{"type":"string","title":"Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Checkin"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/checkins/business/{business_id}":{"get":{"tags":["checkins"],"summary":"Read Checkins By Business","description":"Get checkins for a specific business","operationId":"read_checkins_by_business_api_v1_checkins_business__business_id__get","parameters":[{"name":"business_id","in":"path","required":true,"schema":{"type":"string","title":"Business Id"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Checkin"},"title":"Response Read Checkins By Business Api V1 Checkins Business  Business Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","description":"Root endpoint","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Business":{"properties":{"business_id":{"type":"string","title":"Business Id"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"postal_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Postal Code"},"latitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Latitude"},"longitude":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Longitude"},"stars":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Stars"},"review_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Review Count"},"is_open":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Is Open"},"attributes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Attributes"},"categories":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Categories"},"hours":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Hours"}},"type":"object","required":["business_id"],"title":"Business"},"Checkin":{"properties":{"business_id":{"type":"string","title":"Business Id"},"date":{"type":"string","title":"Date"}},"type":"object","required":["business_id","date"],"title":"Checkin"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Review":{"properties":{"review_id":{"type":"string","title":"Review Id"},"user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id"},"business_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Business Id"},"stars":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Stars"},"useful":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Useful"},"funny":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Funny"},"cool":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Cool"},"text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Text"},"date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Date"},"year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year"},"month":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Month"}},"type":"object","required":["review_id"],"title":"Review"},"Tip":{"properties":{"user_id":{"type":"string","title":"User Id"},"business_id":{"type":"string","title":"Business Id"},"text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Text"},"date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Date"},"compliment_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Compliment Count"},"year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year"}},"type":"object","required":["user_id","business_id"],"title":"Tip"},"User":{"properties":{"user_id":{"type":"string","title":"User Id"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"review_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Review Count"},"yelping_since":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Yelping Since"},"friends":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Friends"},"useful":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Useful"},"funny":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Funny"},"cool":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Cool"},"fans":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Fans"},"elite":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Elite"},"average_stars":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Average Stars"},"compliment_hot":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Compliment Hot"},"compliment_more":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Compliment More"},"compliment_profile":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Compliment Profile"},"compliment_cute":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Compliment Cute"},"compliment_list":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Compliment List"},"compliment_note":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Compliment Note"},"compliment_plain":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Compliment Plain"},"compliment_cool":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Compliment Cool"},"compliment_funny":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Compliment Funny"},"compliment_writer":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Compliment Writer"},"compliment_photos":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Compliment Photos"}},"type":"object","required":["user_id"],"title":"User"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}